+++ ghost-gradle-plugin/src/main/kotlin/com/ghost/gradle/GhostPlugin.kt
@@ -133,7 +133,7 @@ class GhostPlugin : Plugin<Project> {
 
     companion object {
         private const val EXTENSION_NAME = "ghost"
-        private const val DEFAULT_VERSION = "1.1.14"
+        private const val DEFAULT_VERSION = "1.1.16"
 
         private const val PLUGIN_KSP = "com.google.devtools.ksp"
         private const val PLUGIN_KMP = "org.jetbrains.kotlin.multiplatform"